lib:tool:kmod: remove code from the lib that shouldn't be accessible to userspace
