+++ src/kmod/sja1105-clocking.c
@@ -5,10 +5,99 @@
  */
 #include <lib/include/static-config.h>
 #include <lib/include/gtable.h>
-#include <lib/include/clock.h>
 #include <lib/include/spi.h>
 #include <common.h>
 #include "sja1105.h"
+
+/* AGU */
+#define AGU_ADDR    0x100800
+
+struct sja1105_cfg_pad_mii_tx {
+	uint64_t d32_os;
+	uint64_t d32_ipud;
+	uint64_t d10_os;
+	uint64_t d10_ipud;
+	uint64_t ctrl_os;
+	uint64_t ctrl_ipud;
+	uint64_t clk_os;
+	uint64_t clk_ih;
+	uint64_t clk_ipud;
+};
+
+
+/* CGU */
+#define CGU_ADDR    0x100000
+
+/* UM10944 Table 82.
+ * IDIV_0_C to IDIV_4_C control registers
+ * (addr. 10000Bh to 10000Fh) */
+struct sja1105_cgu_idiv {
+	uint64_t clksrc;
+	uint64_t autoblock;
+	uint64_t idiv;
+	uint64_t pd;
+};
+
+/* UM10944 Table 80.
+ * PLL_x_S clock status registers 0 and 1
+ * (address 100007h and 100009h) */
+struct sja1105_cgu_pll_status {
+	uint64_t lock;
+};
+
+/*
+ * PLL_1_C control register
+ *
+ * SJA1105 E/T: UM10944 Table 81 (address 10000Ah)
+ * SJA1105 P/Q/R/S: UM11040 Table 116 (address 10000Ah)
+ *
+ */
+struct sja1105_cgu_pll_control {
+	uint64_t pllclksrc;
+	uint64_t msel;
+	uint64_t nsel; /* Only for P/Q/R/S series */
+	uint64_t autoblock;
+	uint64_t psel;
+	uint64_t direct;
+	uint64_t fbsel;
+	uint64_t p23en; /* Only for P/Q/R/S series */
+	uint64_t bypass;
+	uint64_t pd;
+};
+
+#define CLKSRC_MII0_TX_CLK 0x00
+#define CLKSRC_MII0_RX_CLK 0x01
+#define CLKSRC_MII1_TX_CLK 0x02
+#define CLKSRC_MII1_RX_CLK 0x03
+#define CLKSRC_MII2_TX_CLK 0x04
+#define CLKSRC_MII2_RX_CLK 0x05
+#define CLKSRC_MII3_TX_CLK 0x06
+#define CLKSRC_MII3_RX_CLK 0x07
+#define CLKSRC_MII4_TX_CLK 0x08
+#define CLKSRC_MII4_RX_CLK 0x09
+#define CLKSRC_PLL0        0x0B
+#define CLKSRC_PLL1        0x0E
+#define CLKSRC_IDIV0       0x11
+#define CLKSRC_IDIV1       0x12
+#define CLKSRC_IDIV2       0x13
+#define CLKSRC_IDIV3       0x14
+#define CLKSRC_IDIV4       0x15
+
+/* UM10944 Table 83.
+ * MIIx clock control registers 1 to 30
+ * (addresses 100013h to 100035h) */
+struct sja1105_cgu_mii_control {
+	uint64_t clksrc;
+	uint64_t autoblock;
+	uint64_t pd;
+};
+
+#define XMII_MODE_MAC    0ull
+#define XMII_MODE_PHY    1ull
+#define XMII_SPEED_MII   0ull
+#define XMII_SPEED_RMII  1ull
+#define XMII_SPEED_RGMII 2ull
+#define XMII_SPEED_SGMII 3ull /* Only available for port 4 on R/S */
 
 static void sja1105_cgu_idiv_access(void *buf, struct sja1105_cgu_idiv *idiv,
                                     int write)
