+++ src/kmod/sja1105.h
@@ -156,4 +156,147 @@ int sja1105_device_id_get(struct sja1105_spi_setup *spi_setup,
 /* sja1105-reset.c */
 int sja1105_cold_reset(struct sja1105_spi_setup *spi_setup);
 
+/* sja1105-spi.c */
+int sja1105_static_config_flush(struct sja1105_spi_setup *spi_setup,
+                                struct sja1105_static_config *config);
+
+/* sja1105-status.c */
+
+#define CORE_ADDR   0x000000
+#define ACU_ADDR    0x100800
+
+struct sja1105_general_status {
+	uint64_t configs;
+	uint64_t crcchkl;
+	uint64_t ids;
+	uint64_t crcchkg;
+	uint64_t nslot;
+	uint64_t vlind;
+	uint64_t vlparind;
+	uint64_t vlroutes;
+	uint64_t vlparts;
+	uint64_t macaddl;
+	uint64_t portenf;
+	uint64_t fwds_03h;
+	uint64_t macfds;
+	uint64_t enffds;
+	uint64_t l2busyfds;
+	uint64_t l2busys;
+	uint64_t macaddu;
+	uint64_t macaddhcl;
+	uint64_t vlanidhc;
+	uint64_t hashconfs;
+	uint64_t macaddhcu;
+	uint64_t wpvlanid;
+	uint64_t port_07h;
+	uint64_t vlanbusys;
+	uint64_t wrongports;
+	uint64_t vnotfounds;
+	uint64_t vlid;
+	uint64_t portvl;
+	uint64_t vlnotfound;
+	uint64_t emptys;
+	uint64_t buffers;
+	uint64_t buflwmark; /* Only on P/Q/R/S */
+	uint64_t port_0ah;
+	uint64_t fwds_0ah;
+	uint64_t parts;
+	uint64_t ramparerrl;
+	uint64_t ramparerru;
+};
+
+struct sja1105_port_status_mac {
+	uint64_t n_runt;
+	uint64_t n_soferr;
+	uint64_t n_alignerr;
+	uint64_t n_miierr;
+	uint64_t typeerr;
+	uint64_t sizeerr;
+	uint64_t tctimeout;
+	uint64_t priorerr;
+	uint64_t nomaster;
+	uint64_t memov;
+	uint64_t memerr;
+	uint64_t invtyp;
+	uint64_t intcyov;
+	uint64_t domerr;
+	uint64_t pcfbagdrop;
+	uint64_t spcprior;
+	uint64_t ageprior;
+	uint64_t portdrop;
+	uint64_t lendrop;
+	uint64_t bagdrop;
+	uint64_t policeerr;
+	uint64_t drpnona664err;
+	uint64_t spcerr;
+	uint64_t agedrp;
+};
+
+struct sja1105_port_status_hl1 {
+	uint64_t n_n664err;
+	uint64_t n_vlanerr;
+	uint64_t n_unreleased;
+	uint64_t n_sizerr;
+	uint64_t n_crcerr;
+	uint64_t n_vlnotfound;
+	uint64_t n_ctpolerr;
+	uint64_t n_polerr;
+	uint64_t n_rxfrmsh;
+	uint64_t n_rxfrm;
+	uint64_t n_rxbytesh;
+	uint64_t n_rxbyte;
+	uint64_t n_txfrmsh;
+	uint64_t n_txfrm;
+	uint64_t n_txbytesh;
+	uint64_t n_txbyte;
+};
+
+struct sja1105_port_status_hl2 {
+	uint64_t n_qfull;
+	uint64_t n_part_drop;
+	uint64_t n_egr_disabled;
+	uint64_t n_not_reach;
+	uint64_t qlevel_hwm[8]; /* Only for P/Q/R/S */
+	uint64_t qlevel[8];     /* Only for P/Q/R/S */
+};
+
+struct sja1105_port_status {
+	struct sja1105_port_status_mac mac;
+	struct sja1105_port_status_hl1 hl1;
+	struct sja1105_port_status_hl2 hl2;
+};
+
+struct sja1105_ptp_status {
+	uint64_t syncstate;
+	uint64_t integcy;
+	uint64_t actcorr;
+	uint64_t maxcorr;
+	uint64_t mincorr;
+	uint64_t syncloss;
+	uint64_t locmem;
+	uint64_t minmemvar;
+	uint64_t maxmemvar;
+	uint64_t localsynmem0;
+	uint64_t localsynmem1;
+	uint64_t minasynmem;
+	uint64_t maxasynmem;
+	uint64_t camem;
+	uint64_t synmem;
+	uint64_t moffset[8];
+	uint64_t timer;
+	uint64_t clock;
+};
+int sja1105_port_status_clear(struct sja1105_spi_setup*, int);
+int sja1105_port_status_get_hl1(struct sja1105_spi_setup *spi_setup,
+                                struct sja1105_port_status_hl1 *status,
+                                int port);
+int sja1105_port_status_get(struct sja1105_spi_setup*,
+                            struct sja1105_port_status*,
+                            int port);
+int  sja1105_general_status_get(struct sja1105_spi_setup*,
+                                struct sja1105_general_status*);
+void sja1105_general_status_show(struct sja1105_general_status*,
+                                 char*, size_t,
+                                 uint64_t device_id);
+
 #endif
